# 🚀 Make IP validation 2x faster