--- a/utils/ips.go
+++ b/utils/ips.go
@@ -2,7 +2,7 @@ package utils
 
 import "net"
 
-// IsIPv4 is plagiarism of net.ParseIP,
+// IsIPv4 works the same way as net.ParseIP,
 // but without check for IPv6 case and without returning net.IP slice, whereby IsIPv4 makes no allocations.
 func IsIPv4(s string) bool {
 	for i := 0; i < net.IPv4len; i++ {
@@ -33,14 +33,10 @@ func IsIPv4(s string) bool {
 		s = s[ci:]
 	}
 
-	if len(s) != 0 {
-		return false
-	}
-
-	return true
+	return len(s) == 0
 }
 
-// IsIPv6 is plagiarism of net.ParseIP,
+// IsIPv6 works the same way as net.ParseIP,
 // but without check for IPv4 case and without returning net.IP slice, whereby IsIPv6 makes no allocations.
 func IsIPv6(s string) bool {
 	ellipsis := -1 // position of ellipsis in ip