@efectn 2022-10-17
Can you create a PR for https://github.com/gofiber/utils